requirements.txt
torch==2.1.0
torchvision==0.15.0
timm==0.6.11
streamlit==1.21.0
Pillow==9.0.1
facenet-pytorch==2.5.2
mediapipe==0.9.0
opencv-python==4.6.0.66
scikit-learn==1.0.2
numpy==1.24.2
